zynqmp_pmufw: Fix violations for MISRA Rule 10_4
This patch fix following violation
misra_c_2012_rule_10_4_violation: Both operands of an operator in
which the arithmetic conversions are performed shall have
the same essential type category.

To solve this MISRA violation do explicit type cast as per requirement.


Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_core.c
@@ -846,7 +846,7 @@ static void PmFpgaLoad(const PmMaster *const master,
     Status = XFpga_PL_BitStream_Load(&XFpgaInstance, BitStreamAddr,
 				     KeyAddr, Flags);
 
-    if ((XST_SUCCESS == Status) && ((Flags & XFPGA_AUTHENTICATION_OCM_EN) ==
+    if ((XST_SUCCESS == (s32)Status) && ((Flags & XFPGA_AUTHENTICATION_OCM_EN) ==
 		XFPGA_AUTHENTICATION_OCM_EN)) {
 	FSBL_Store_Restore_Info.IsOCM_Used = TRUE;
     }
